pyproject.toml
[project]
name = "project-netease"
version = "0.1.0"
description = "An interactive story game with AI-generated content using LangChain and Flask"
readme = "README.md"
authors = [
    {name = "Ke Xu", email = "kexu567@gmail.com"},
    {name = "Fangzhan Lin", email = "xxx@gmail.com"},
]
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.1.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "flask>=3.1.1",
    "flask-socketio>=5.5.1",
    "python-socketio>=5.13.0",
    "python-engineio>=4.12.1",
]

[project.scripts]
game = "run_game:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=5.0.0",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.uv]
dev-dependencies = [] 